# ota_dashboard_web/openapi.json
{
  "components": {
    "schemas": {
...
      "CriticalBannerResponse": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "raised_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raised At"
          },
          "severity": {
            "anyOf": [
              {
                "enum": [
                  "info",
                  "warn",
                  "error",
                  "critical"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "active"
        ],
        "title": "CriticalBannerResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KnobValue": {
        "additionalProperties": false,
        "properties": {
          "default": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "KnobValue",
        "type": "object"
      },
      "KnobsUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "knobs": {
            "additionalProperties": true,
            "title": "Knobs",
            "type": "object"
          }
        },
        "title": "KnobsUpdateRequest",
        "type": "object"
      },
      "KnobsUpdateResponse": {
        "additionalProperties": false,
        "properties": {
          "applied": {
            "additionalProperties": true,
            "title": "Applied",
            "type": "object"
          },
          "routine_id": {
            "title": "Routine Id",
            "type": "string"
          }
        },
        "required": [
          "routine_id",
          "applied"
        ],
        "title": "KnobsUpdateResponse",
        "type": "object"
      },
      "RaiseBannerRequest": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "severity": {
            "enum": [
              "info",
              "warn",
              "error",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "title"
        ],
        "title": "RaiseBannerRequest",
        "type": "object"
      },
      "RoutineKnobsResponse": {
        "additionalProperties": false,
        "properties": {
          "knobs": {
            "items": {
              "$ref": "#/components/schemas/KnobValue"
            },
            "title": "Knobs",
            "type": "array"
          },
          "routine_id": {
            "title": "Routine Id",
            "type": "string"
          }
        },
        "required": [
          "routine_id",
          "knobs"
        ],
        "title": "RoutineKnobsResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WhyEntryModel": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "kind",
          "description",
          "payload"
        ],
        "title": "WhyEntryModel",
        "type": "object"
      },
      "WhyResponse": {
        "additionalProperties": false,
        "properties": {
          "email_id": {
            "title": "Email Id",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/WhyEntryModel"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "email_id",
          "entries"
        ],
        "title": "WhyResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Operator dashboard for the OTA framework. Read-only audit, approval queue, /why drill-down, routine knob editor.",
    "title": "OTA Dashboard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/notifications/banner": {
      "delete": {
        "operationId": "clear_banner_api_v1_notifications_banner_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CriticalBannerResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Clear Banner",
        "tags": [
          "notifications"
        ]
      },
      "get": {
        "operationId": "get_banner_api_v1_notifications_banner_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CriticalBannerResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Banner",
        "tags": [
          "notifications"
        ]
      },
      "post": {
        "operationId": "raise_banner_api_v1_notifications_banner_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RaiseBannerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CriticalBannerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Raise Banner",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/routines/{routine_id}/knobs": {
      "get": {
        "operationId": "get_knobs_api_v1_routines__routine_id__knobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineKnobsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Knobs",
        "tags": [
          "knobs"
        ]
      },
      "post": {
        "operationId": "update_knobs_api_v1_routines__routine_id__knobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnobsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnobsUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Knobs",
        "tags": [
          "knobs"
        ]
      }
    },
    "/api/v1/why/{email_id}": {
      "get": {
        "operationId": "why_api_v1_why__email_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Why",
        "tags": [
          "why"
        ]
      }
    },
...
  }
}